backend/requirements.txt
asgiref==3.8.1
boto3==1.20.26
botocore==1.23.54
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==41.0.7
dj-database-url==2.1.0
Django==5.1.4
django-anymail==13.0.1
django-ckeditor-5==0.2.18
django-cors-headers==3.14.0
django-jazzmin==3.0.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
environs==10.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jmespath==0.10.0
marshmallow==3.20.1
moviepy==2.2.1
packaging==23.2
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.21
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.5.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==7.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
